AppliedJobs dropdown works

+++ my_app/dashboard/applications/applied_jobs/frontend/src/components/body/utils/jobs_list/DeleteJob.js
@@ -1,17 +1,29 @@
 //DeleteJob.js
 
-export const deleteJob = async (jobId)=> {
+export const deleteJob = async (jobId, jobCompany, jobTitle)=> {
 	try{
-		const response = await fetch('/api/delete-job', {
-      			method: 'DELETE',
-      			headers: { 'Content-Type': 'application/json' },
-      			credentials: 'include',  // Sends Flask session cookie
-			body: JSON.stringify({ jobId })
-    		});
+	
+		const confirmDelete = window.confirm(`Are you sure you want to delete ${jobTitle} at ${jobCompany}?`);
 
-    		if (!response.ok) {
-     			throw new Error(`HTTP error! status: ${response.status}`);
-    		}
+		if (confirmDelete){
+			const response = await fetch('/api/delete-job', {
+                        	method: 'DELETE',
+                        	headers: { 'Content-Type': 'application/json' },
+                        	credentials: 'include',  // Sends Flask session cookie
+                        	body: JSON.stringify({ jobId })
+                	});
+
+               		const result = await response.json();
+
+                	if (response.ok) {
+                        	console.log('Success:', result.message);
+                	}
+                	else {
+                        	console.error('Server error:', result.error);
+                	}
+
+		}
+
 	}
 
 	catch(error){
